requirements.txt
#langchain core
langchain
langchain-core

#open ai integration
langchain-openai
openai

#Anthropic integration
langchain-anthropic


#google gemini (PaLM) integration
langchain-google-genai
google-generativeai

#hugging face integration
langchain-huggingface
transformers
huggingface-hub

#environment variable management
python-dotenv

#Machine learning utilities
numpy
scikit-learn

streamlit
